+++ ui/src/main/java/com/kotcrab/vis/ui/widget/HighlightTextArea.java
@@ -144,6 +144,7 @@ public class HighlightTextArea extends ScrollableTextArea {
 		}
 
 		layoutPool.free(layout);
+		updateScrollLayout();
 	}
 
 	@Override