+++ ui/CHANGES.md
@@ -2,6 +2,9 @@
 - **Fixed**: `Spinner.TextFieldEventPolicy` is now public (was package-private)
 - **Improved**: [#220](https://github.com/kotcrab/vis-editor/issues/220) when sub menu can't fit on the right side of parent menu, it will be shown on the side that has more available space (before in such case it was always shown on the left side)
 - **Improved**: When mouse is moved from sub-menu to parent menu, selection of menu item in sub-menu will be removed.
+- **Skin changes**:
+   - **Added new icons**: `icon-maximize`, `icon-minimize`, `icon-restore`, `icon-close-titlebar`
+   - **Added style**: `VisImageButtonStyle`: `close-titlebar`
 
 #### Version: 1.2.3 (LibGDX 1.9.4)
 - **Added**: constructor `LinkLabel (CharSequence text, CharSequence url, LinkLabelStyle style)`