+++ ui/CHANGES.md
@@ -3,6 +3,7 @@
 - **Fixed**: `HighlightTextArea` scroll pane not immediately updated after changing text using `setText()`
 - **Improved**: [#220](https://github.com/kotcrab/vis-editor/issues/220) when sub menu can't fit on the right side of parent menu, it will be shown on the side that has more available space (before in such case it was always shown on the left side)
 - **Improved**: When mouse is moved from sub-menu to parent menu, selection of menu item in sub-menu will be removed.
+- **Improved**: [#222](https://github.com/kotcrab/vis-editor/issues/222) Added clipping to BusyBar
 - **Skin changes**:
    - **Added new icons**: `icon-maximize`, `icon-minimize`, `icon-restore`, `icon-close-titlebar`
    - **Added style**: `VisImageButtonStyle`: `close-titlebar`